Move `codeElementPositions` to common utils
